Fix cyclic imports reported by pylint.
I moved function_name from nengo.utils.connection to
nengo.utils.functions, which seems to be a better place to me
because it is more related to function in general than
specifically connections. In particular, nengo.builder.operator
was using it in SimPyFunc which is not related to connections,
but nodes.